requirements.txt
pandas
numpy
nltk
ufal.udpipe
tqdm
gensim
umap-learn
hdbscan
scikit-learn